streamlit_app/requirements.txt
streamlit
Pillow
streamlit_drawable_canvas
bentoml
opencv-python